% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{lllmr_status}
\alias{lllmr_status}
\title{Check Ollama Status}
\usage{
lllmr_status(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical. If TRUE, suppresses messages.}
}
\value{
Logical indicating whether Ollama is reachable.
}
\description{
Check Ollama Status
}
