% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{run_server}
\alias{run_server}
\title{Run the httpuv server (blocking) — called in a background process}
\usage{
run_server(app_dir, model, host, port)
}
\description{
Run the httpuv server (blocking) — called in a background process
}
\keyword{internal}
